requirements.txt
flake8==6.0.0
cocotb==1.8.0
cocotbext-spi==0.3.2
numpy==1.26.4
pytest==8.1.1